Split BlueSkyPost component into two
