+++ __tests__/helpers/fetchers/BlueskyFetcher.test.ts
@@ -6,9 +6,9 @@ import { BlueskyFetcher } from "#/screens/Home/fetchers/BlueskyFetcher";
 import "#tests/mocks/commonMocks";
 
 // Mock the BskyPost component to avoid complex dependencies
-jest.mock("#/components/posts/BlueskyPost", () => ({
+jest.mock("#/components/posts/BlueskyPostCard", () => ({
   __esModule: true,
-  default: "BskyPost",
+  default: "BlueskyPostCard",
 }));
 
 describe("BlueskyFetcher", () => {
